Move test_value
